fix: Remove NULLS FIRST/LAST in more MySQL Window functions
